feat: rename to repl
since this is now a repl don't use the name tutorial anymore
just a simple repl :)
